+++ CHANGES.md
@@ -10,6 +10,7 @@
   * Bugfix: Reset delay after connection is successful
   * Feature: yapic.json parses strings to datetimes automatically, no longer need to rely on Pandas for datetime parsing
   * Bugfix: #491 - dictionary resized during iteration in ByBit
+  * Bugfix: #494 - added status argument to liquidations callback
 
 ### 1.9.0 (2021-04-25)
   * Bugfix: Fix Binance subscriptions when subscribing to more than one candle